move go files to proper directories
