+++ .github/workflows/ci.yml
@@ -21,10 +21,10 @@ jobs:
         go-version: 1.x
 
     - name: Build
-      run: cd scripts && go build main.go
+      run: go build
 
     - name: Test
-      run: cd scripts && go test ./...
+      run: go test . ./internal/... ./subcmd/...
 
   diff:
     name: 'Compare Pages and Site'
