+++ internal/slacklog/channel.go
