+++ internal/slacklog/converter.go
