+++ internal/slacklog/emoji.go
