+++ internal/slacklog/filetype.go
