+++ internal/slacklog/generator.go
