+++ internal/slacklog/message.go
