+++ internal/slacklog/store.go
