+++ internal/slacklog/testdata/downloader/vim-jp.png
