+++ internal/slacklog/thread.go
