+++ subcmd/convert_exported_logs.go
@@ -14,7 +14,7 @@ import (
 	"os"
 	"path/filepath"
 
-	slacklog "github.com/vim-jp/slacklog/lib"
+	"github.com/vim-jp/slacklog-generator/internal/slacklog"
 )
 
 func ConvertExportedLogs(args []string) error {
